+++ src/instance/TempFileSystem.h
@@ -129,7 +129,7 @@ public:
 	// NodeIndexPool
 	//
 	// Lock-free pool of node index numbers
-	IndexPool<intptr_t>	NodeIndexPool;
+	IndexPool<int64_t> NodeIndexPool;
 
 	// MaximumNodes
 	//
@@ -313,7 +313,7 @@ private:
 		// index
 		//
 		// The node index value
-		intptr_t const index;
+		int64_t const index;
 
 		// mode
 		//
@@ -601,8 +601,8 @@ private:
 		// Index (VirtualMachine::Node)
 		//
 		// Gets the node index within the file system (inode number)
-		__declspec(property(get=getIndex)) intptr_t Index;
-		virtual intptr_t getIndex(void) const override;
+		__declspec(property(get=getIndex)) int64_t Index;
+		virtual int64_t getIndex(void) const override;
 
 		// Mode (VirtualMachine::Node)
 		//
